unify timestamp generation in one generator instead of three
shows a nice feature of the constructor-based init
